requirements.txt
selenium
pandas
xlsx2csv
pillow
pytesseract
torch
tensorflow
transformers
opencv-python